client/pirate_watcher/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
watchdog==4.0.2